requirements.txt
invoke
tqdm
pyyaml
transformers
sentencepiece

--index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio